requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.2
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
install-jdk==1.1.0
sqlparse==0.5.3
